Handle annotated assignment in 3.6
In 3.6, annotated assignment was added, which takes the form of `a: int = 0`.
This would cause errors, since `:` is an operator, and its child would contain
nodes of type `annassign`.

We handle these the same way that we handle normal assignments, but by dropping
into the assignment logic at an offset in the expression we're handling.