Fix self parameter for mutated class methods
Previously, mutated class methods were
always killed, because no `self` arg
was passed to their calls and thus
raised a TypeError for invalid number
of arguments.

The class methods in the mutants dict are
stored without reference to any class instance
and therefore not bound to any instance.
We need to pass the `self` arg explicitly
in those cases.

Note that the call to the original method
is a bound method (the method is looked up
via the `self` parameter), thus we do not
pass the `self` arg in this case.