Supply status for missing exit codes
Following on from the work in "Handle -9 PyTest return code",
discussion with @Otto-AA and @boxed revealed it would be desirable
to handle missing return codes in full generality.

The simplest way I could think of to _do_ that, trying to eliminate
side effects, is to convert status_by_exit_code to a defaultdict,
with a default value of "suspicious" - one already defined, and
handled, elsewhere in the codebase.