Handle 1-element mutant_names tuples with spaces
As at the status quo, calling mutmut run with two or more mutant
specs in quotes (such as "mutant_1 mutant_2") blew up with an
AssertionError in collect_source_file_mutation_data, stating
"Filtered for specific mutants, but nothing matches. ...."

This commit handles that case, by splitting the tuple's element
and replacing the tuple if the resulting split has 2 or more
elements.