Fix unchanged string mutants (#437)
* Add test for github issue 435

* Filter out identical string mutants with different values

As at the status quo, operator_string took the prefix of the node's
value, trimming that off the rest to leave an approximation.

This worked well enough until some wombat (hi) fed in a regex
string consisting only of escape sequences and whitespace.  For
whatever reason, the node represented that value as r'(rest of text...)
which the status quo dutifully trimmed to (rest of text...).

As the two text strings didn't match, status quo dutifully let the
mutant through despite it being a no-op.

This commit adds an additional check to catch the _unmodified_ node
value being equal to the mutated node value, skipping those
candidate mutants.